Mitigations for potential SNAT port exhaustion
There's a possibility (to be confirmed) that our flakiness with E2E tests may be caused by SNAT port exhaustion in the Actions runners. This commit includes improvements with recommendations that could mitigate that:

1. Ensure that HTTP clients are re-used and not recreated every time
2. Ensure that responses' `Body` streams are always closed properly, so that the client is not idle
