+++ tests/e2e/runtime/runtime_test.go
@@ -53,9 +53,9 @@ func getAPIResponse(t *testing.T, testName, runtimeExternalURL string) (*daprAPI
 
 	resp, err := utils.HTTPGetRaw(url)
 	require.NoError(t, err)
+	defer resp.Body.Close()
 	require.Equal(t, resp.StatusCode, http.StatusOK)
 
-	defer resp.Body.Close()
 	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return nil, err
